refactor(core): extract the sensitive-scan concern from _hooklib (#321, slice 1/4)
_hooklib.py was 1,263 lines and 46 functions serving 28 of 44 consumers. First of
four planned slices; it takes the module to 1,049.

_sensitivelib owns the H-09b/H-10b sensitive-line scan as ONE concern: the crypto
and secret detectors, the pinned security diff argv, the path-aware diff walk,
and the digests that bind a recorded gate pass to the exact lines it reviewed.
_pathnorm holds norm_path as the dependency FLOOR - two lines in its own module
on purpose, because every path classifier normalizes separators, so whichever
cluster moved first would otherwise have to import from _hooklib (a cycle, since
_hooklib re-exports it) or carry its own copy.

THE SEAM WAS MEASURED, NOT GUESSED. An AST pass showed the cluster referenced
exactly ONE symbol from the rest of _hooklib, and that NOTHING in the rest of
_hooklib referenced the cluster. A one-way edge with no back-reference is what
made this safe to move without touching a consumer.

They belong together because they fail together: #279 showed a greedy diff-header
parse un-exempting an unrelated file, and an unpinned prefix config silently
re-opening it. Splitting them would let one move without the others.

_hooklib re-exports every moved name, so all 59 consuming files are UNCHANGED and
parity rests on pre-existing hook tests that did not move. The public surface is
preserved exactly: nothing gained, and the only losses are _PLUS_B_PREFIX /
_PLUS_DEV_NULL - both private, both with zero references outside the code that
moved with them.

Cleared H-09b through the auth-crypto reviewer, which verified rather than took
my word: it SHA-256'd all 12 moved symbols against origin/main (identical),
re-ran the #279 exploit end-to-end in a throwaway repo (still blocks), proved
fail-CLOSED by deleting the new modules with a real violation staged, and found
no import cycle across the 50-module local graph. Its two LOW findings are fixed
here: the facade had split a comment from the `_HOST = None` it documents, and
`import hashlib` was left dead once the digests moved out.

THREE FINDINGS FROM DOING IT:

1. My own "13 unused public symbols" claim was WRONG, and I checked before
   deleting. Twelve have real internal uses - implementation detail that merely
   lacks an underscore - and reset_host is live in two files under
   plugins/ca/hooks/tests/, a tree my first AST scan never globbed. Only
   migration_globs is genuinely unreferenced. There was no dead code to delete.

2. test_taskwrite.py hardcoded a list it called _hooklib's "complete stdlib-only
   dependency closure", and this change made that false. Its subprocesses died on
   ModuleNotFoundError, surfacing as "process 1 never acquired the real lock" - a
   lock diagnosis for an import failure. Confirmed against clean main (1185 OK
   there) rather than assumed. The closure is now DERIVED by walking imports, so
   slices 2-4 cannot reintroduce it. Assertions untouched.

3. Seven test_colorlib palette tests fail in MY checkout and not in CI. They are
   not caused by this change: proven by copying these modules into a clean
   worktree, where all 30 pass. They render the LIVE statusline and require all
   seven custom palette colours to appear, which depends on .codearbiter/ counts -
   and the security-pass run above appended audit rows. Pre-existing fragility,
   filed separately rather than worked around here.

sync-core globs core/pysrc/*.py, so both new modules vendored byte-identically
into all three plugins with no tool change: 50 core files x 3 plugins verified,
and the reviewer confirmed all four copies of each file share one blob ID.

ca 2.10.4 -> 2.10.5, ca-codex 0.3.3 -> 0.3.4, ca-pi 0.1.39 -> 0.1.40: new files
under plugins/*/hooks/ are shipped payload for all three.
