+++ .codearbiter/gate-events.log
@@ -1130,3 +1130,10 @@ Claude-Session: https://claude.ai/code/session_015ZDVU1BzUqnnPVHbZ397bX') stages
 [2026-07-29T16:00:06Z] REMIND [H-09] host=claude hook=post-write-edit.py | Crypto/TLS pattern detected. Run the crypto-compliance check + dispatch auth-crypto-reviewer (no MD5/SHA1/DES/3DES/RC2/RC4/Blowfish; do not disable TLS verification). The commit will block until the gate records a pass.
 [2026-07-30T03:02:19Z] BLOCK [H-03] host=claude hook=pre-bash.py | 'git add -A' / 'git add .' / 'git add --all' / 'git add -u' are prohibited. Stage files explicitly (commit-gate skill).
 [2026-07-30T03:02:47Z] REMIND [H-12] host=claude hook=post-write-edit.py | docs/codex-parity-testing.md is governed by ADR-0006-broad-adoption-oss-posture (Broad-adoption OSS posture, optimizing for adoption over a commercial vertical). If this change contradicts it, route to /ca:reconcile or /ca:adr — do not drift silently.
+[2026-07-30T11:14:58Z] BLOCK [H-03] host=codex hook=pre-bash.py | 'git add -A' / 'git add .' / 'git add --all' / 'git add -u' are prohibited. Stage files explicitly (commit-gate skill).
+[2026-07-30T11:18:55Z] REMIND [H-12] host=claude hook=post-write-edit.py | plugins/ca/hooks/tests/test_taskwrite.py is governed by ADR-0004-database-free-stdlib-only-architecture (Database-free architecture; Python hooks are stdlib-only). If this change contradicts it, route to /ca:reconcile or /ca:adr — do not drift silently.
+[2026-07-30T11:23:55Z] BLOCK [H-09b] host=claude hook=pre-bash.py | This commit introduces crypto/TLS changes, but no security-gate pass is recorded (.codearbiter/.markers/security-gate-passed). Run the crypto-compliance gate (it records the pass), then commit. To bypass a security gate, /override requires its heavier security-acknowledgement path.
+[2026-07-30T11:31:38Z] BLOCK [H-03] host=claude hook=pre-bash.py | 'git add -A' / 'git add .' / 'git add --all' / 'git add -u' are prohibited. Stage files explicitly (commit-gate skill).
+[2026-07-30T11:31:49Z] BLOCK [H-18] host=claude hook=pre-bash.py | .codearbiter/CONTEXT.md is the activation switch every enforcement hook reads (#159) — shell rewrites, edits, or deletions that could flip `arbiter: enabled` off or corrupt its frontmatter are prohibited. Edit it through the sanctioned init path.
+[2026-07-30T11:33:50Z] BLOCK [H-19] host=claude hook=pre-bash.py | The .codearbiter/.markers/ security-gate-passed / migration-gate-passed tokens are recorded only by the sanctioned gate producers (#160) — a shell redirect, write verb, or interpreter invocation (python/node/perl/ruby/sh) naming a gate marker forges a security/migration gate pass and is prohibited.
+[2026-07-30T11:34:13Z] REMIND [H-07] host=codex hook=post-write-edit.py | Dependency manifest changed. Dispatch dependency-reviewer before committing (ORCHESTRATOR §5).
