+++ .codearbiter/overrides.log
@@ -21,3 +21,4 @@
 [2026-07-16T17:05:59.9905219-04:00] | BY: SUaDtL@users.noreply.github.com | GATE: H-18 protected CONTEXT.md write | REASON: Task 12 requires a prose-only Pi host vocabulary update; preserve and verify the 34-byte activation frontmatter at SHA-256 519c223415be8cd907ca2d696f56314996b8aa46170d029ff6bbbef2fe7b8bcd before and after the immediate edit
 [2026-07-16T17:11:47.8577605-04:00] | BY: SUaDtL@users.noreply.github.com | GATE: H-05 audit-log line-ending repair | REASON: the immediately prior H-18 override append used PowerShell CRLF in this LF-only change; remove only its terminal carriage-return byte so git diff --check passes while preserving every audit character and record
 2026-07-19T21:45:48.4804940-04:00 | BY: SUaDtL@users.noreply.github.com | GATE: append-only sprint-log byte rewrite | REASON: Content-preserving CRLF-to-LF normalization of the already appended sprint records is required for the mandatory git diff --check gate; exact text and order remain unchanged; user explicitly approved this immediate action.
+2026-07-30T07:22:18.5387323-04:00 | BY: SUaDtL@users.noreply.github.com | DEV: enter | NOTE: —
